SAM-3149 Hibernate 4 fixes for samigo (#3929)
- AuthzQueriesFacade.removeAuthorizationByAgentQualifierAndFunction
- QuestionPoolFacadeQueries.getAllPoolsWithAccess
- QuestionPoolFacadeQueries.getSubPoolSize
- QuestionPoolFacadeQueries.getSubPoolSizes
- QuestionPoolFacadeQueries.getCountItemFacades

+++ samigo/samigo-services/src/java/org/sakaiproject/tool/assessment/shared/impl/questionpool/QuestionPoolServiceImpl.java
@@ -198,7 +198,7 @@ public class QuestionPoolServiceImpl
   /**
    * Get the size of a subpool.
    */
-  public int getSubPoolSize(Long poolId)
+  public long getSubPoolSize(Long poolId)
   {
     try
     {
